requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
scipy
numpy
lightfm
scikit-learn
joblib
python-dotenv
asyncpg
databases